Introduce base class for Trafikverket camera
