+++ homeassistant/components/trafikverket_camera/sensor.py
@@ -13,13 +13,12 @@ from homeassistant.components.sensor import (
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import DEGREE
 from homeassistant.core import HomeAssistant, callback
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import StateType
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import DOMAIN
 from .coordinator import CameraData, TVDataUpdateCoordinator
+from .entity import TrafikverketCameraNonCameraEntity
 
 PARALLEL_UPDATES = 0
 
@@ -92,39 +91,26 @@ async def async_setup_entry(
 
     coordinator: TVDataUpdateCoordinator = hass.data[DOMAIN][entry.entry_id]
     async_add_entities(
-        TrafikverketCameraSensor(coordinator, entry.entry_id, entry.title, description)
+        TrafikverketCameraSensor(coordinator, entry.entry_id, description)
         for description in SENSOR_TYPES
     )
 
 
-class TrafikverketCameraSensor(
-    CoordinatorEntity[TVDataUpdateCoordinator], SensorEntity
-):
+class TrafikverketCameraSensor(TrafikverketCameraNonCameraEntity, SensorEntity):
     """Representation of a Trafikverket Camera Sensor."""
 
     entity_description: TVCameraSensorEntityDescription
-    _attr_has_entity_name = True
 
     def __init__(
         self,
         coordinator: TVDataUpdateCoordinator,
         entry_id: str,
-        name: str,
-        entity_description: TVCameraSensorEntityDescription,
+        description: TVCameraSensorEntityDescription,
     ) -> None:
         """Initiate Trafikverket Camera Sensor."""
-        super().__init__(coordinator)
-        self.entity_description = entity_description
-        self._attr_unique_id = f"{entry_id}-{entity_description.key}"
-        self._attr_device_info = DeviceInfo(
-            entry_type=DeviceEntryType.SERVICE,
-            identifiers={(DOMAIN, entry_id)},
-            manufacturer="Trafikverket",
-            model="v1.0",
-            name=name,
-            configuration_url="https://api.trafikinfo.trafikverket.se/",
-        )
-        self._update_attr()
+        self.entity_description = description
+        super().__init__(coordinator, entry_id)
+        self._attr_unique_id = f"{entry_id}-{description.key}"
 
     @callback
     def _update_attr(self) -> None:
@@ -132,8 +118,3 @@ class TrafikverketCameraSensor(
         self._attr_native_value = self.entity_description.value_fn(
             self.coordinator.data
         )
-
-    @callback
-    def _handle_coordinator_update(self) -> None:
-        self._update_attr()
-        return super()._handle_coordinator_update()
